cree el evento de integracion para desvincular la empresa y el user. Defini para cuando se realiza una eliminacion se pueda validar si es el unico permiso que tiene el usuario en la Empresa, tambien sea desvinculado.

+++ identity/src/main/java/com/kynsof/identity/application/command/userPermisionBusiness/deleteRelation/DeleteUserBusinessRelationCommandHandler.java
@@ -2,27 +2,33 @@ package com.kynsof.identity.application.command.userPermisionBusiness.deleteRela
 
 import com.kynsof.identity.domain.dto.UserPermissionBusinessDto;
 import com.kynsof.identity.domain.interfaces.service.IUserPermissionBusinessService;
+import com.kynsof.identity.infrastructure.services.kafka.producer.userBusiness.ProducerDeleteUserBusinessRelationEventService;
 import com.kynsof.share.core.domain.bus.command.ICommandHandler;
 import java.util.List;
 import org.springframework.stereotype.Component;
-
 
 @Component
 public class DeleteUserBusinessRelationCommandHandler implements ICommandHandler<DeleteUserBusinessRelationCommand> {
 
     private final IUserPermissionBusinessService service;
+    private final ProducerDeleteUserBusinessRelationEventService deleteUserBusinessRelationEventService;
 
-    public DeleteUserBusinessRelationCommandHandler(IUserPermissionBusinessService service) {
+    public DeleteUserBusinessRelationCommandHandler(IUserPermissionBusinessService service,
+            ProducerDeleteUserBusinessRelationEventService deleteUserBusinessRelationEventService) {
         this.service = service;
+        this.deleteUserBusinessRelationEventService = deleteUserBusinessRelationEventService;
     }
 
     @Override
     public void handle(DeleteUserBusinessRelationCommand command) {
         // Recuperar el estado actual
         List<UserPermissionBusinessDto> currentPermissions = service.findByUserAndBusiness(command.getUserId(),
-                        command.getBusinessId())
+                command.getBusinessId())
                 .stream()
                 .toList();
-        this.service.delete(currentPermissions);
+        if (!currentPermissions.isEmpty()) {
+            this.service.delete(currentPermissions);
+            this.deleteUserBusinessRelationEventService.delete(currentPermissions.get(0).getUser().getId(), currentPermissions.get(0).getBusiness().getId());
+        }
     }
 }
